Show triangulation in rooms window
